Implement Login & Registry system

+++ Server/scheduling.football.games/src/main/java/com/proiect/isi/scheduling/football/games/controllers/LocationController.java
@@ -1,7 +1,9 @@
 package com.proiect.isi.scheduling.football.games.controllers;
 
 import com.proiect.isi.scheduling.football.games.dto.LocationDto;
+import com.proiect.isi.scheduling.football.games.services.AuthService;
 import com.proiect.isi.scheduling.football.games.services.LocationService;
+import jakarta.servlet.http.Cookie;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -17,34 +19,54 @@ import java.util.UUID;
 public class LocationController {
 
     private final LocationService locationService;
+    private final AuthService authService;
 
     @PostMapping("/locations")
-    public ResponseEntity<?> postLocation(@Valid @RequestBody LocationDto locationDto){
+    public ResponseEntity<?> postLocation(@Valid @RequestBody LocationDto locationDto,
+                                            @CookieValue(name = "AuthToken") String cookie){
 
-        UUID locationId = locationService.addLocation(locationDto);
+        if(cookie.equals("Admin")) {
+            UUID locationId = locationService.addLocation(locationDto);
+
+            return ResponseEntity
+                    .status(HttpStatus.OK)
+                    .body(locationId);
+        }
 
         return ResponseEntity
-                .status(HttpStatus.OK)
-                .body(locationId);
+                .status(HttpStatus.UNAUTHORIZED)
+                .build();
     }
 
     @GetMapping("/locations")
-    public ResponseEntity<?> getLocations(){
+    public ResponseEntity<?> getLocations(@CookieValue(name = "AuthToken") String cookie){
 
-        List<LocationDto> locations = locationService.getAllLocations();
+        if(cookie.equals("Admin") || cookie.equals("Player")) {
+            List<LocationDto> locations = locationService.getAllLocations();
+
+            return ResponseEntity
+                    .status(HttpStatus.OK)
+                    .body(locations);
+        }
 
         return ResponseEntity
-                .status(HttpStatus.OK)
-                .body(locations);
+                .status(HttpStatus.UNAUTHORIZED)
+                .build();
     }
 
     @DeleteMapping("/locations")
-    public ResponseEntity<?> deleteLocation(@RequestParam UUID id){
+    public ResponseEntity<?> deleteLocation(@RequestParam UUID id, @CookieValue(name = "AuthToken") String cookie){
 
-        locationService.deleteLocation(id);
+        if(cookie.equals("Admin")) {
+            locationService.deleteLocation(id);
+
+            return ResponseEntity
+                    .status(HttpStatus.OK)
+                    .build();
+        }
 
         return ResponseEntity
-                .status(HttpStatus.OK)
+                .status(HttpStatus.UNAUTHORIZED)
                 .build();
     }
