+++ Server/scheduling.football.games/src/main/java/com/proiect/isi/scheduling/football/games/controllers/MatchController.java
@@ -21,42 +21,58 @@ public class MatchController {
     private final MatchService matchService;
 
     @PostMapping("/matches")
-    public ResponseEntity<?> postMatch(@Valid @RequestBody MatchDto MatchDto){
+    public ResponseEntity<?> postMatch(@Valid @RequestBody MatchDto MatchDto, @CookieValue(name = "AuthToken") String cookie){
 
-        UUID matchId = matchService.addMatch(MatchDto);
+        if(cookie.equals("Admin") || cookie.equals("Player")) {
+            UUID matchId = matchService.addMatch(MatchDto);
 
+            return ResponseEntity
+                    .status(HttpStatus.OK)
+                    .body(matchId);
+        }
         return ResponseEntity
-                .status(HttpStatus.OK)
-                .body(matchId);
+                .status(HttpStatus.UNAUTHORIZED)
+                .build();
     }
 
 
     @GetMapping("/matches")
     public ResponseEntity<?> getMatchesByDateRange(
             @RequestParam(value = "startDate", required = false) @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") Optional<Date> startDate,
-            @RequestParam(value = "endDate", required = false) @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") Optional<Date> endDate) {
+            @RequestParam(value = "endDate", required = false) @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") Optional<Date> endDate,
+            @CookieValue(name = "AuthToken") String cookie) {
 
-        List<MatchDto> matches = new ArrayList<>();
-        if(startDate.isEmpty() && endDate.isEmpty()){
-            matches = matchService.getAllMatches();
+        if(cookie.equals("Admin") || cookie.equals("Player")) {
+            List<MatchDto> matches = new ArrayList<>();
+            if (startDate.isEmpty() && endDate.isEmpty()) {
+                matches = matchService.getAllMatches();
+            }
+
+            if (startDate.isPresent() || endDate.isPresent()) {
+                matches = matchService.getMatchesByDateRange(startDate, endDate);
+            }
+
+            return ResponseEntity
+                    .status(HttpStatus.OK)
+                    .body(matches);
         }
-
-        if(startDate.isPresent() || endDate.isPresent()) {
-            matches = matchService.getMatchesByDateRange(startDate, endDate);
-        }
-
         return ResponseEntity
-                .status(HttpStatus.OK)
-                .body(matches);
+                .status(HttpStatus.UNAUTHORIZED)
+                .build();
     }
 
 
     @DeleteMapping("/matches")
-    public ResponseEntity<?> deleteMatch(@RequestParam UUID id){
-        matchService.deleteMatch(id);
+    public ResponseEntity<?> deleteMatch(@RequestParam UUID id, @CookieValue(name = "AuthToken") String cookie){
+        if(cookie.equals("Admin")) {
+            matchService.deleteMatch(id);
 
+            return ResponseEntity
+                    .status(HttpStatus.OK)
+                    .build();
+        }
         return ResponseEntity
-                .status(HttpStatus.OK)
+                .status(HttpStatus.UNAUTHORIZED)
                 .build();
     }
 }
